feat(storage): hot-swap safeties for removable media
Closes the last remainder of the roadmap's wear-leveling / log-offload item.

An SD card is a connector, so it can be pulled mid-write, and the failure is a
quiet one: the driver still reports a mounted volume, writes fail into nothing,
and code that does not check the return carries on believing it has storage.
Six subsystems here touch fs::FS directly with nothing tracking whether the
medium is still present.

services/hotswap is one state machine per volume (ABSENT / READY / FAULTED).
dws_hotswap_ready() is the fail-closed gate callers check before any filesystem
call and dws_hotswap_io(ok) reports every outcome; a run of
DWS_HOTSWAP_FAIL_THRESHOLD consecutive errors declares the medium gone and
unmounts it at once so nothing writes through a stale handle, and a probe every
DWS_HOTSWAP_PROBE_MS remounts a card that comes back.

The threshold is the point: one failed write is not proof of removal (a
transient bus error, a full volume) and tearing down a working mount over it
would be its own bug, so any success resets the run. A volume also starts
ABSENT rather than READY, and present-but-unmountable stays ABSENT.

Core is pure and takes an explicit now, so it is host-tested with a synthetic
clock (20 cases incl. counter saturation, rollover-safe probe pacing, and the
boot-time underflow when begin() runs a few ms after reset).

HW-verified on a P4 with a real card: three good writes left it ready/faults:0,
an unmount underneath the app made the next three writes really fail and the
third faulted it, the following write was refused with "storage not ready"
rather than attempted, and the probe then remounted it by itself (mounts 1->2).

Also clears three SonarCloud findings that were fallout from recent commits,
all behavior-preserving: dead range-parsing left behind when http_delivery's
duplicate range core was removed (plus its now-unused string.h), a redundant
cast in wifi_sniffer, and the host-side multicast validator, which is extracted
into parse_mcast_group() so the group copy is bounded by a length the function
just proved instead of an unbounded strncpy relying on a distant check.

+++ src/services/wifi_sniffer/wifi_sniffer.cpp
@@ -119,7 +119,7 @@ bool dws_wifi_scan_due(const WifiScan *s, uint32_t now_ms)
     if (!s)
         return false;
     // Unsigned subtraction is correct across a millis() rollover.
-    return (uint32_t)(now_ms - s->last_hop_ms) >= s->dwell_ms;
+    return (now_ms - s->last_hop_ms) >= s->dwell_ms;
 }
 
 uint8_t dws_wifi_scan_next(WifiScan *s, uint32_t now_ms)
